LayoutTests/http/tests/loading/unfinished-load-back-to-cached-page-callbacks-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didFailLoadWithError
main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishLoadForFrame
This test dumps frame load callbacks. It is only useful inside of WebKitTestRunner.

This link goes to a page with a slow loading subresource. Bug 119416 - Going "back" to a cached page from a page with outstanding subresource loads breaks scrolling, amongst other issues.

In the broken case, the second page gets a didFinishLoad callback intertwined with the restoration of the cached page, even though it's already gotten a didFailLoad callback.
The final 4 callbacks look like:
didFailLoadWithError
didStartProvisionalLoadForFrame
didFinishLoadForFrame
didCommitLoadForFrame

When fixed, the final 4 callbacks should be:
didFailLoadWithError
didStartProvisionalLoadForFrame
didCommitLoadForFrame
didFinishLoadForFrame
